Minimal robustness fixes in SysUtil.java and XCom.java

+++ src/jxm/xb/XCom.java
@@ -138,7 +138,7 @@ public class XCom {
         { return new DoubleRef(value); }
 
         public static DoubleRef valueOf(final Double value)
-        { return new DoubleRef( value.intValue() ); }
+        { return new DoubleRef( value.doubleValue() ); }
 
         public static DoubleRef valueOf(final String value)
         { return new DoubleRef( Double.parseDouble(value) ); }
@@ -384,6 +384,7 @@ public class XCom {
             return true;
         }
 
+        // BUGNOTE: value can be null (VariableStore default constructor sets value = null) or empty; value.charAt(0) throws NullPointerException or StringIndexOutOfBoundsException; fix: add null and empty check before charAt
         public boolean isVar()
         { return !constant && ( value.charAt(0) == '$' || isWritableSVarSCut(value) ); }
 
@@ -2522,6 +2523,7 @@ public class XCom {
     public static String sqStringEllipsis(final String str, final int maxLen)
     {
         return sqStringEllipsis(
+            // BUGNOTE: when maxLen < 5, maxLen - 5 is negative and substring throws StringIndexOutOfBoundsException; fix: add maxLen < 5 to the short-circuit condition
             ( str == null || str.length() < maxLen - 2 ) ? str : ( str.substring(0, maxLen - 5) + "..." )
         );
     }
@@ -2840,6 +2842,7 @@ public class XCom {
                     while( (i < len - 1) && (nch >= '0') && (nch <= '7') ) {
                         ++i;
                         code.append(nch);
+                        // BUGNOTE: after ++i, i + 1 can equal len, making str.charAt(i + 1) throw StringIndexOutOfBoundsException; fix: guard with (i + 1 < len) ? str.charAt(i + 1) : '\0'
                         nch = str.charAt(i + 1);
                     }
                     sb.append( (char) Integer.parseInt( code.toString(), 8 ) );
@@ -3037,6 +3040,7 @@ public class XCom {
     {
         if( sep.equals("") ) {
             for(final VariableStore item : parts) {
+                // BUGNOTE: item.value can be null (VariableStore default constructor); calling codePoints() or split() on null throws NullPointerException; fix: skip items where item.value is null
                 for( final int cp : item.value.codePoints().toArray() ) retVal.add( new VariableStore( true, new String( Character.toChars(cp) ) ) );
               //for( final char ch : item.value.toCharArray() ) retVal.add( new VariableStore( true, String.valueOf(ch) ) );
             }
@@ -3620,6 +3624,7 @@ public class XCom {
         for(int i = 0; i < str.length() && idx < 2; ++i) {
             if( str.charAt(i) == '\\' && i + 5 < str.length() && str.charAt(i + 1) == 'u' ) {
                 // Extract the hexadecimal value and convert it into a normal character
+                // BUGNOTE: Integer.parseInt throws NumberFormatException if the four hex digits are not valid hex; fix: wrap in try/catch and treat the backslash as a literal character on failure
                 res[idx++] = (char) Integer.parseInt( str.substring(i + 2, i + 6), 16 );
                 i += 5;
             }
